requirements.txt
# Dependencies for Moltbook Bot
requests==2.31.0
python-dotenv==1.0.0